Fix a few utilities

+++ codec/wavelet.cpp
@@ -52,14 +52,14 @@ size_t ForwardWavelet1D(const int16_t *src, int16_t *dst, size_t len) {
   const size_t mid_pt = len - (len / 2);
 
   // Do the odd coefficients first
-  for (int i = 1; i < len; i += 2) {
+  for (int i = 1; i < static_cast<int>(len); i += 2) {
     int next = NormalizeIndex(i + 1, len);
     int prev = NormalizeIndex(i - 1, len);
     dst[mid_pt + i / 2] = src[i] - (src[prev] + src[next]) / 2;
   }
 
   // Do the even coefficients second
-  for (int i = 0; i < len; i += 2) {
+  for (int i = 0; i < static_cast<int>(len); i += 2) {
     int next = static_cast<int>(mid_pt) + NormalizeIndex(i + 1, len) / 2;
     int prev = static_cast<int>(mid_pt) + NormalizeIndex(i - 1, len) / 2;
     dst[i / 2] = src[i] + (dst[prev] + dst[next] + 2) / 4;
@@ -85,14 +85,14 @@ void InverseWavelet1D(const int16_t *src, int16_t *dst, size_t len) {
   const size_t mid_pt = len - (len / 2);
 
   // Do the even coefficients first
-  for (int i = 0; i < len; i += 2) {
+  for (int i = 0; i < static_cast<int>(len); i += 2) {
     int prev = static_cast<int>(mid_pt) + NormalizeIndex(i - 1, len) / 2;
     int next = static_cast<int>(mid_pt) + NormalizeIndex(i + 1, len) / 2;
     dst[i] = src[i / 2] - (src[prev] + src[next] + 2) / 4;
   }
 
   // Do the odd coefficients second
-  for (int i = 1; i < len; i += 2) {
+  for (int i = 1; i < static_cast<int>(len); i += 2) {
     int src_idx = static_cast<int>(mid_pt) + i / 2;
     int prev = NormalizeIndex(i - 1, len);
     int next = NormalizeIndex(i + 1, len);
